Chore: Extract `frontMatter` logic into a separate file
